style(design): Update site colors and base css variables
Update colors and color variables in main.scss. Now uses a pallet with 4
colors and a base and alternative colors based on the pallets colors.
Add _utils.scss with utility functions and mixins. Update all other
files that used the old color variables to the new variables. Remove
grey color varibles and content-width.